requirements.txt
Flask
spacy
pdfminer.six
python-docx
